+++ test/util/components/ResetCacheProvider.tsx
@@ -1,0 +1,10 @@
+import { SWRConfig } from 'swr'
+import * as React from 'react'
+
+export function ResetCacheProvider({ children }) {
+  return (
+    <SWRConfig value={{ dedupingInterval: 0, provider: () => new Map() }}>
+      {children}
+    </SWRConfig>
+  )
+}
